+++ broker/hook/grpc.go
@@ -1,7 +1,7 @@
 // Copyright (c) Abstract Machines
 // SPDX-License-Identifier: Apache-2.0
 
-package hookcallout
+package hook
 
 import (
 	"context"
